Fix data leak in groups.md
I'm pretty sure that putting e-mailadresses of students on a public Github repository is in violation with the Dutch law. This incident should also be reported to the AP (Autoriteit Persoonsgegevens) and this pull request should not be merged, but history should be rewritten and the file should be changed so it doesn't include e-mail adresses and force pushed.